requirements.txt
# Google Cloud AI and Agent Development Kit dependencies
google-cloud-aiplatform>=1.35.0
google-cloud-dialogflow>=2.24.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Core dependencies for agent development
openai>=1.0.0
langchain>=0.1.0
langchain-google-vertexai>=1.0.0

# Web framework for agent endpoints
fastapi>=0.104.0
uvicorn>=0.24.0

# Data handling and utilities
pydantic>=2.4.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Logging and monitoring
structlog>=23.1.0
rich>=13.6.0

# Optional: Jupyter for experimentation
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: Additional AI/ML libraries
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0